Added support for Google Gemini

+++ src/Zatomic.AI.Providers/Anthropic/AnthropicChatRequest.cs
@@ -33,7 +33,7 @@ namespace Zatomic.AI.Providers.Anthropic
 		public AnthropicChatThinking Thinking { get; set; }
 
 		[JsonProperty("top_k", NullValueHandling = NullValueHandling.Ignore)]
-		public float? TopK { get; set; }
+		public int? TopK { get; set; }
 
 		[JsonProperty("top_p", NullValueHandling = NullValueHandling.Ignore)]
 		public float? TopP { get; set; }
