+++ src/Zatomic.AI.Providers/Hyperbolic/HyperbolicChatRequest.cs
@@ -33,7 +33,7 @@ namespace Zatomic.AI.Providers.Hyperbolic
 		public float? Temperature { get; set; }
 
 		[JsonProperty("top_k", NullValueHandling = NullValueHandling.Ignore)]
-		public float? TopK { get; set; }
+		public int? TopK { get; set; }
 
 		[JsonProperty("top_p", NullValueHandling = NullValueHandling.Ignore)]
 		public float? TopP { get; set; }
