+++ src/Zatomic.AI.Providers/Meta/MetaChatRequest.cs
@@ -27,7 +27,7 @@ namespace Zatomic.AI.Providers.Meta
 		public float? Temperature { get; set; }
 
 		[JsonProperty("top_k", NullValueHandling = NullValueHandling.Ignore)]
-		public float? TopK { get; set; }
+		public int? TopK { get; set; }
 
 		[JsonProperty("top_p", NullValueHandling = NullValueHandling.Ignore)]
 		public float? TopP { get; set; }
